LayoutTests/sputnik/Conformance/12_Statement/12.1_Block/S12.1_A4_T1-expected.txt

CONSOLE MESSAGE: line 80: SyntaxError: Unexpected token '}'
S12.1_A4_T1

PASS Expected parsing failure

TEST COMPLETE
